chore(clean-types): moves config and serialisationtypes to own file
