+++ runem/command_line.py
@@ -9,7 +9,8 @@ from runem.informative_dict import InformativeDict
 from runem.log import error, log
 from runem.runem_version import get_runem_version
 from runem.types_wip.common import JobNames
-from runem.types_wip.types import OptionConfig, OptionsWritable
+from runem.types_wip.runem_config import OptionConfig
+from runem.types_wip.types import OptionsWritable
 from runem.utils import printable_set
