+++ runem/job.py
@@ -1,7 +1,8 @@
 import typing
 
 from runem.types_wip.common import FilePathList, JobTags
-from runem.types_wip.types import FilePathListLookup, JobConfig, JobWhen
+from runem.types_wip.runem_config import JobConfig, JobWhen
+from runem.types_wip.types import FilePathListLookup
 
 
 class NoJobName(ValueError):
