+++ runem/job_execute.py
@@ -13,7 +13,8 @@ from runem.job import Job
 from runem.job_wrapper import get_job_wrapper
 from runem.log import error, log
 from runem.types_wip.common import FilePathList, JobTags
-from runem.types_wip.types import FilePathListLookup, JobConfig
+from runem.types_wip.runem_config import JobConfig
+from runem.types_wip.types import FilePathListLookup
 from runem.types_wip.types_jobs import (
     HookSpecificKwargs,
     JobFunction,
