+++ runem/job_filter.py
@@ -5,7 +5,7 @@ from runem.config_metadata import ConfigMetadata
 from runem.job import Job
 from runem.log import log
 from runem.types_wip.common import JobNames, JobPhases, JobTags, PhaseName
-from runem.types_wip.types import JobConfig, PhaseGroupedJobs
+from runem.types_wip.runem_config import JobConfig, PhaseGroupedJobs
 from runem.utils import printable_set
