+++ runem/job_runner_simple_command.py
@@ -2,7 +2,7 @@ import shlex
 import typing
 
 from runem.run_command import run_command
-from runem.types_wip.types import JobConfig
+from runem.types_wip.runem_config import JobConfig
 
 
 def validate_simple_command(command_string: str) -> typing.List[str]:
