+++ runem/job_wrapper_python.py
@@ -4,7 +4,7 @@ from importlib.util import module_from_spec
 from importlib.util import spec_from_file_location as module_spec_from_file_location
 
 from runem.types_wip.errors import FunctionNotFound
-from runem.types_wip.types import JobWrapper
+from runem.types_wip.runem_config import JobWrapper
 from runem.types_wip.types_jobs import JobFunction
