+++ tests/test_runem.py
@@ -27,11 +27,9 @@ from runem.runem import (
     _update_progress,
     timed_main,
 )
-from runem.types_wip.types import (
+from runem.types_wip.runem_config import (
     Config,
-    FilePathListLookup,
     GlobalSerialisedConfig,
-    HookName,
     HookSerialisedConfig,
     JobConfig,
     Jobs,
@@ -39,6 +37,7 @@ from runem.types_wip.types import (
     PhaseGroupedJobs,
     UserConfigMetadata,
 )
+from runem.types_wip.types import FilePathListLookup, HookName
 from runem.types_wip.types_jobs import JobReturn, JobTiming
 from tests.intentional_test_error import IntentionalTestError
 from tests.sanitise_reports_footer import sanitise_reports_footer
